requirements.txt
appnope==0.1.3
asgiref==3.6.0
backcall==0.2.0
click==8.1.3
colorama==0.4.6
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==4.12.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
Pygments==2.14.0
PyJWT==2.7.0
PyMySQL==1.0.2
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
pyzmq==25.0.2
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.3
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
